+++ lib/sparql/query/result/formats/json_decoder.ex
@@ -46,12 +46,12 @@ defmodule SPARQL.Query.Result.JSON.Decoder do
   defp decode_value(%{"type" => "literal", "value" => value, "xml:lang" => language}),
     do: RDF.Literal.new(value, language: language)
 
+  defp decode_value(%{"type" => type, "value" => value, "datatype" => datatype})
+       when type in ~w[literal typed-literal],
+       do: RDF.Literal.new(value, datatype: datatype)
+
   defp decode_value(%{"type" => "literal", "value" => value}),
     do: RDF.Literal.new(value)
-
-  defp decode_value(%{"type" => type, "value" => value, "datatype" => datatype})
-    when type in ~w[literal typed-literal],
-    do: RDF.Literal.new(value, datatype: datatype)
 
   defp decode_value(%{"type" => "bnode", "value" => value}),
     do: RDF.BlankNode.new(value)