+++ CHANGELOG.md
@@ -7,10 +7,14 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 
 ## Unreleased
 
-### Changed
+### Fixed
 
-- recognize more values in CSV results as IRIs (only some very common URI schemes
-  are recognized still; if there's really need, this requires further work)
+- The JSON query result decoder didn't recognize the datatype of literals when the JSON
+  value object was `"type" : "literal"`. Only for those with `"typed-literal"` the `"datatype"` 
+  field was interpreted.    
+- Not all IRI values in CSV results were recognized as IRIs. Values starting with the most  
+  common URI schemes (http/https, urn, ftp, file, ldap, mailto, geo, data) are now recognized 
+  correctly. 
 
 [Compare v0.3.5...HEAD](https://github.com/rdf-elixir/sparql-ex/compare/v0.3.5...HEAD)